build: Add support for direnv (#2484)
[direnv](https://direnv.net/) is a nice utility. This PR proposes a very
simple integration where the Composer bins of the project are added to
your path. As a result, you can now type:

```shell
$ php-cs-fixer
$ phpunit
```

And they will work. Additionally, it will automatically load a
`.envrc.local` (which is not committed) if present, allowing one to
further extend their own envrc.

Note that this is not required by anyone to use, but can allow one to
customize their own experience with this project.