test: Introduce WebmozartsStrictPHPUnit (#2487)
This PHPUnit extension aims at solving the issue described in #2486,
i.e. ensure that objects are strictly compared rather than loosely
compared.

It is not a universal solution, for instance four our traces we still
need this `TraceAssertions` as `Trace` is an interface and when we want
to compare traces we likely want to compare the state without any
potential laziness, but it still removes any need for boilerplate for
plain value objects.