perf(Tracing): Provide a better source for the Tracing benchmarks (#2476)
The current sources shipped are not very practical. It is hard to update
and results in a ton of console warnings.

The issue with this sort of thing is that a live project is much better,
but then you need to deal with all sort of dependencies. A bigger
project is better, but a bigger project often also has higher
infrastructure (DB/extensions, etc.) requirements.

As discussed together, I created
https://github.com/infection/benchmark-source which should provide a
sizeable codebase we can use for benchmark purposes.

With this PR, the tracing profiling now works. There is more
changes/refinement to do, but we now get a Blackfire profile which looks
like this:

<img width="1489" height="939" alt="Screenshot 2025-10-26 at 18 16 04"
src="https://github.com/user-attachments/assets/2e5ea61e-5080-4348-ab53-ffafb1fe0d4b"
/>