chore: Fix various benchmark configuration (#2496)
This fixes various occurrences of the new benchmark traces introduced in
#2476.

I decided to keep the old entries just in case someone has them locally
to not cause any surprise. It would be better for that person to delete
them, but I also understand the frustration to have to deal with it when
it is something you do not care about.