test: Make BenchmarkTest more resilient
Although the test is not actually executed in the e2e tests, the PHPUnit
providers are still.

The issue with `Safe\realpath()` here is that:

- If it fails, like it did in
  https://github.com/infection/infection/actions/runs/18856716771/job/53806284939?pr=2497,
then it result in a test failure even if the test is belonging to an
excluded group (i.e. is not to be executed).
- Even if we used `realpath()` instead of the `Safe` variant, we would
  end up with `false` instead of the desired path, which will make the
failure message more criptic than necessary.