test: Silence output when dumping the autoloader for the benchmark (#2500)
Because we have the following in our `composer.json`:

```json
    "autoload-dev": {
        "psr-4": {
            "Infection\\Benchmark\\": "tests/benchmark",
```

The following code:

- `tests/benchmark/MutationGenerator/sources/`
- `tests/benchmark/Tracing/benchmark-source/`

is included. As a result, when we execute `composer dump
--classmap-authoritative` as part of the benchmark scripts, we get a
wall of warnings as the code in those directories do not follow the
classmap (hence is skipped).

I can't find a nice way to address this without trading off convenience
(we do not want to have to manually include every single source code
file under `tests/benchmark`). Hence, my proposal is to add `--quiet` to
silence those warnings.