test(benchmark): Update the benchmark scripts (#2499)
I do not like too much this sort of PR that does a lot of various
things, but untangling it is too much effort for what it is, both for me
and any reviewer.

This PR:

- Removes the `generate-mutations.php` and `provide-traces.php` scripts.
Those were originally meant as alternatives to the respective
`profile.php` files. However, I think #2460 is a more elegant solution
and thus it has since then be made irrelevant.
- Adjust the handling of the `$maxCount`: for both benchmarks it is now
a `positive-int`. This is to have consistency on both side and simplify
a bit the downstream code.
- Make the max count argument into an option. The rationale is that
otherwise, doing `php profile.php -1` will fail, as there is no option
`-1`. Instead you should call `php profile.php "-1"` but I think at this
point using an option is better.
- Add a bit more validation & user friendly messages upon failure for
the max count option.
- Bumped the maximum count for the tracing benchmark: we generate ~250
traces, which is relatively fast hence I do not think we need to limit
it by default (one can still, for debugging purposes or whatever).
- Bumped the maximum number of mutations. I was thinking of removing the
limit (like for the tracing), but we generate over 36k mutations...
which does take a while.
- Rename `BenchmarkTest` for `BenchmarkSmokeTest` with an explanation.
This is to both make it more obvious what this test is about and not
confusing future incoming `*Bench` tests (will be detailed in a
dedicated PR as they get introduced).
- Update `BenchmarkTest` to use the respective `profile.php` files
instead.
- Rename `--max-mutations-count` to `--max-mutation-count` for
consistency & correctness.
- Fix `provide-traces-closure.php` max loop check, previously if called
with maxCount=2, then it would generate only 1 trace. As a result
calling it with maxCount=1 resulted in an error as no trace would be
generated.