test(benchmark): Test the output of the profile scripts (#2514)
As shown in #2513, #2499 and #2476, the current profile scripts are
somewhat fragile and too little tested. They are not meant to be a
critical piece of code we need to babysit but they were probably broken
for years without us noticing.

This PR also adds the output to the tested subject, which gives a bit
more guarantee without breaking a leg.