refactor(Str): Introduce line ending replacement utilities (#2522)
Extracted from #2504 for better visibility.

Currently we have:

- `normalizeLineReturn()` which replaces the line endings of the value
by the unix line endings.
- `LineReturnNormalizer::normalize()` which replaces the line endings of
the value by the system line endings. Note that the implementation was
flawed in that if it received a non-unix line endings, it would not
replace it by the system line ending.

The initial proposition for the replacement was:

- `normalizeLineReturn()` -> `Str::toLinuxLineReturn()`.
- `LineReturnNormalizer::normalize()` -> `Str::toSystemLineReturn()`.

I decided to change "linux" for "unix", which is more universal and
"line return" for "line endings" which is more commonly used (and for
example is the term used by PHPUnit). So the final replacement is:

- `normalizeLineReturn()` -> `Str::toUnixLineEndings()`.
- `LineReturnNormalizer::normalize()` -> `Str::toSystemLineEndings()`.

Both implementations have been consolidated and tests were added.