refactor(Str): consolidate the remaining lines normalization utilities (#2523)
It seems I was too hasty in #2504 and made some incorrect assumptions.
So this PR limits its scope a great deal and only does the following:

- Replaces `normalize_trailing_spaces()` by `Str::rTrimLines()`. Unlike
the original proposal of naming it `normalize()` or `trimLines()`, I
added the `r` prefix since we are doing an `rtrim()` and not a `trim()`.
- Renamed `Str::trimLineReturns()` into `::removeOuterBlankLines()`
which better describes what it does.
- Add a lot more tests.

Further improvements will be done separately.