fix(Str): Make `::removeOuterBlankLines()` convert line endings (#2527)
While working on #2504 I noticed that `Str::removeOuterBlankLines()`
(previously `Str::trimLineReturns()`) replaces the line ending by the
system line ending instead of the unix one.

I think this behaviour is a bit confusing, especially since
`Str::rTrimLines()` (previously `normalize_trailing_spaces()`) does not.
I can't find a reason to wish for this behaviour and the CI does not
appear to mind either.

So this PR proposes to normalize the line ending as well for consistency
(and I think it's a more intuitive behaviour).