refactor(Str): Rework `::removeOuterBlankLines()` (#2531)
Closes #2528.

As discussed in the linked issue, in terms of usage it makes little
sense for `::removeOuterBlankLines()` to not trimming the trailing
whitespaces of lines.

However, this may also be a bit confusing as the method name suggests no
such thing. As discussed, we decided to go with the name
`::cleanForDisplay()` which is not the best description for _what it
does_, but at least better conveys the _intent_.