chore(benchmark): Fix usage of the deprecrated Blackfire `--samples` option (#2517)
The option `--samples` is now a noop option. To leverage samples we are
now expected to do it ourselves.

In our case it is not that simple, because we need to re-do the setup
step for every execution of the profiled script. Indeed, the profiled
code is a non rewindable generator.

As a result, in this PR, the closure script no longer does the setup +
return the function to profile, but instead returns a closure doing
that.