test(benchmark): Enusure any lazy state is executed (#2544)
When I first created #2452 I was testing against some uncommitted work
where the trace state was lazily evaluated, hence just iterating over
the traces was not doing much, a lot more computation was only triggered
once accessing the trace test.

I thought this would apply to the existing benchmark, but it appears the
`JUnitTestExecutionInfoAdder`, which is executed by
`CoveredTraceProvider`, load any lazy data to do its own thing. In other
words, once the provider yields a trace, the state of the trace is fully
initialised.

As a result, whilst I think this PR still has value because it has the
merit of making it explicit, it does not result in any chance in the
profile.

For good measure I also checked for the mutation but we do not have any
lazy state there.

This PR is the last of #2452, which closes a 3 weeks journey :)