test(Builder): Add more builders (#2539)
Originally was #2488 but in there I tackled `ConfigurationFactoryTest`,
which is quite a beast and I would like to handle separately.

So this PR is a continuation of #2482. As a reminder, the goal is:

- Remove the need of mocks for `@final` value objects.
- Remove boilerplate to compare value objects.
- Make the value objects `final`.

Unless I missed something, this PR enables #2483.

Note that although this adds quite a bit of code, it is mostly 1) the
builders themselves, for which the implementation is indeed verbose 2)
assertions + builders tests. The tests code that is updated to use
builders is otherwise simpler.