ci: Move the benchmark test to its dedicated workflow (#2502)
`BenchmarkSmokeTest` was configured as an integration test but always
incomplete as its sources were never provided in the CI.

To execute the benchmark test, we need the sources, and I do not think
it is worth troubling ourselves with the whole matrix we have for our
unit & integration tests. So I propose to move it to its own workflow.

Another benefit is if we add actual benchmarks (as opposed to scripts to
execute with Blackfire), then we can include them there without
impacting the rest of the CI. This is already sketched out in
https://github.com/infection/infection/pull/2516.