refactor: Improve the source filtering readability (#2542)
This PR essentially moves the whole filter + git filter + git options
block to a separate method. Then a bit more stuff is extracted for
readability purposes.

I believe on its own it helps a bit to follow what is going on there. My
main motivation behind this is to introduce a value object to make a lot
simpler to navigate through that piece of code (see #2395).