test: Simplify MutantTest
I was orginally planning to remove this test (this is an extracted piece
from #2488). But in the end, it's a case where the entity is not as
trivial. I don't think it holds much value, but I do not mind as much
especially with the reduce boilerplate of the builder.