refactor(SafeDOMXPath): Use the XML namespace (#2559)
While working on #2543, I noticed that `DOMDocument::load(string
$filename)` is a bit faster than getting the file content and using
`DOMDocument::loadXML(string $xml)`.

The main issue is... namespaces. When loading the string, we can,
beforehand, remove the namespaces found. This allows us to do queries
without caring about them at all.

However, if you load the document from the file, then this solution is
out of the window. You need to register a namespace and then uglify all
of your queries with it.

So this PR focuses on switching to namespaces. Switching to
`DOMDocument::load()` will be done separately.

<hr/>

You can find the results of the benchmark in #2561.

Personally, I hate either approach: I find ugly to load the file to
remove the namespaces, but the namspaces are a b**** to handle and makes
queries ugly.

So either we think the performance benefit is worth the tradeoff and we
merge this PR, or we don't in which case I'll adapt #2543 to stick to
loading the file first and stripping namespaces from the string.