refactor(ClassName): Consolidate utilities into a ClassName utility (#2592)
While working on #2591 I noticed several things:

- We have a need to get a class short name in multiple places. I already
noticed it in `FilterBuilder` in the past but at that time it was a bit
overkill or I was a bit undecided about introducing a utility class.
- We have `SourceTestClassName` which is internal but now is part of the
source code as it may end up being used (indirectly, via the
`BaseMutatorTest`) for custom mutators.
- in #2591 I need to extend the beahviour of `SourceTestClassName`.

This PR proposes to consolidate all of thus under a `ClassName` utility.
It:

- Introduces `::getShortName()` and leverages it.
- Renames `SourceTestClassName` to `ClassName`.
- Renames `::getSourceClassName()` to `::getCanonicalSourceClassName()`
and likewise for test. It may confuse briefly what "canonical" refers to
in this context, but the fact is that this is not a super all purpose
generic utility: it is quite specific to infection and our naming
conventions. Hence, canonical here refers to our "normally expected
value following our conventions". I'll also update the documentation in
an upcoming PR.