test(FileSystem): Automatically move to the created tmp directory (#2583)
`FileSystemTestCase` test a temporary directory that is usable for the
time of the test and removed afterwards. I think it is a bit more
intuitive that by default, we are located in that temporary directory.
As a result, if you do:

```php
touch('index.xml'),
```

Then it is automatically created in that temporary directory and not the
current working directory.

This change also highlighted the fact that two tests were specifically
relying on the current working directory being the project root, so I
made this explicit.