refactor(SafeDOMXPath): Introduce a more friendly API to SafeDOMXpath (#2585)
Extracted from #2543.

From #2598 it is apparent that PHPStan does not account correctly for
`DOMNodeList` array access (`$domNodeList[0] ≠ $domNodeList->item(0)`)
and only the later form shows the type issues. As a result of this,
making our code stricter makes it unnecessarily ugly.

This PR introduces two key changes:

- We no longer rely on the magic getter in `SafeDOMXPath`. Instead we
make `SafeDOMXpath#document` a public readonly property.
- Instead of having untyped `::query()` that may return anything, I
introduced a couple of utility methods that makes it a bit easier to
read (at the cost of more heavy testing of `SafeDOMXPath` – although IMO
they have the merit of showing a bit more the API of this service for
those unfamiliar with XPath).

This PR is a strict refactoring, no change of behaviour is intended
here.