refactor(report): Make the report default location behaviour more consistent (#2602)
Extracted from #2588.

For the `JUnitReportLocator` report, the default location is handled
directly by the `Container` whereas for the `IndexXmlCoverageLocator`
the default location is handled in its constructor.

This PR proposes to:

- Move the "figure out the default location" to the respective report,
i.e. `JUnitReportLocator` will provide the default location instead of
the `Container`.
- Introduce a `::create()` method for both locators: in this create
method. This offers several advantages:
- We can simplify the state of `__construct()`. This allows to more
easily leverage promoted properties and make the class readonly
(eventually).
- This makes it more flexible: if we want to use a specific values for
the locator, in tests for example, we can do so more easily as we now
have an API that offers no transformation of the input.
- The more boring logic can be handled in this method, namely
canonicalizing the paths and using a default location fallback if none
is provided (which is the case most for the source code at least).

Note that default location is exposed by the method
`::getDefaultLocation()`. The name and choice to use a method is in
preparation of #2588.