test(report): Consolidate the Index & JUnit report locator tests (#2603)
This PR is a preparatory step for #2588.

It adds some tests and rename other bits so that the names align with
the tests in #2588. The goal is to reduce the noise there and make it
easier to identify how the tests have been moved around. It is otherwise
too hard to follow if there is a change of behaviour or where did what
test go.