test(FileSystem): Update the make_tmp_dir function (#2584)
I'm not too happy with the current implementation of `make_tmp_dir`:

- It is a plain function which is a bit annoying with the autoloading. I
managed to move all the other functions that was in this `Helpers.php`,
this one is the last one.
- The implementation is a bit confusing and while it works it's really
tailored to an approach, which is relying on the `TEST_TOKEN` and having
one main tmp directory, in `FileSystemTestCase`.
- I do not think the approach of having a "namespace" directory makes
much sense:
- It is is more likely to break things with parallelisation. We have to
use `TEST_TOKEN` to actively avoid that, but that's again very specific
and if one launches the tests differently it breaks.
- We do not use consistently this in the project anyway, so that the
assumption that all temporary files created by the Infection test suite
are in one directory that can be safely deleted is false.

Instead the proposed solution is – I hope, a bit more straightforward:

- The utility function `make_tmp_dir` is moved to `FS::tmpDir()` (the
naming is picked after other Symfony Filesystem method names, e.g.
`readLink()`, `readFile()`.
- The implementation of `::tmpDir()` is closer to the native
implementation of `tempnam()`.
- `FileSystemTestCase` now creates a temporary directory in the system
default temporary path based on the used class name. It still has a
random suffix (which uses `bin2hex(random_bytes())` rather than
`random_int`) so we keep the best of both worlds: the names are still
namespaced, making it easy to remove leftover files generated by
Infection tests (shouldn't be needed in practice though) but remains
sufficiently unique that it is unlikely to collide with the temporary
directory created from another thread.
- Just for good measure I also introduced `::tmpFile()`, which is
basically `tempnam()` renamed for consistency.