refactor: Introduce a ReportLocator interface for the coverage report locators (#2588)
Extracted from #2543.

As of now, the JUnit & Index locator are 95% the same code and same
tests, yet different classes. More annoying as well is that those two
classes are needed elsewhere hence we directly mock them.

This PR introduces a `ReportLocator` interface. This has several
benefits:

- This makes mocking simpler (no more not final `@final` class)
- It introduces a `FakeLocator` implementation that is otherwise a bit
more annoying to mimic with mocks.
- It introduces a `FixedLocator` which is useful when we know exactly
where the report is, i.e. there is no lookup to do. Quite useful for
some tests but there is nothing test specific in the implementation.
- It introduces a `BaseReportLocator` base implementation*.

*: I had a lot of back and forth on how to proceed there. My conclusion
is that we have the same behaviour for both locators. What differs is
the exceptions because we want to fine-tune the exception message. So
this base implementation offers the hooks to do so.

To have a finer control on the FileSystem, new methods have been
introduced. I refrained from letting it creep up to other places for now
however. Note that the implementation
is quite rudimentary, but I believe `fidry/filesystem` is more adequate
for a more proper implementation. We don't really have a need for this
yet so I just kept the naïve implementation for now.

I reverted the original chance to move the caching to a dedicated
locator (`MemoizedLocator`) as there is currently issues in the
container to achieve this. See
https://github.com/sanmai/di-container/issues/25