refactor(git): Remove dead code (#2614)
If we are in a git diff, then the base branch is already fetched at the
time the `Configuration` object is created. It needs to be nullable
however because it will be `null` if no git filter is provided.

This will hopefully be fixed soonish by a better design (still drafting
#2381), but meanwhile this removes any potential confusion (at least in
the context of where this is needed/used).