fix(filter): Add the git base branch to the Configuration (#2619)
Vaguely similar to #2617.

The git base branch can be explicitly passed by the user but we have a
fallback, namely `GitDiffFileProvider::provideDefaultBase()`.

However, for some reasons, although we fetch its values in the
`ConfigurationFactory`, we did not pass it to `Configuration` unless the
user explicitly set it.

Prior to #2614, this was not a problem other than some potential
confusion and unnecessary calls made (which should not have any
performance impact since we cached the result).

So in a way this is only really a fix because of #2614, but I think it
was incorrect before that too (it just didn't cause a hard failure to
the user).