fix(filter): Fix isForGitFilter not always set correctly (#2617)
## Context

The user may apply a git filter in two different ways:

- `--git-diff-filter=AD`
- `--git-diff-lines` which is an alias for `--git-diff-filter=AM`.

Now the issue is that since this is two distinct options, we need to
distinguish the two inputs and validate them accordingly as they are
mutually exclusive.

I think there was some confusions before. The code evolved a lot since
the filters were first introduced (#1368, #1632, #2394, #2542). Checking
that the right options are passed correctly is now clearly enforced in
`RunCommand` (see #2542) and `ConfigurationFactory` responsibility is to
pass that `$sourceFilesFilter: string` value and the git related values
when a git filter is applied.


## Problem

Right now `Configuration#isForGitDiff` is set only if the user passed
`--git-diff-lines`. This is incorrect as the goal of this value is to
convey that we are applying a git diff of sorts for:

- `IndexXmlCoverageParser` to throw a more understandable exception to
the user when no coverage is found.
- `FileMutationGenerator/NodeMutationGenerator` to only mutate the nodes
that are part of the diff.

In other words, it should not matter which option was used.


## Solution

This PR corrects the value.