feat(git): Use the full branch name (#2621)
A git branch name has two forms:

- full path: `refs/remotes/origin/HEAD`
- shorthand: `origin/HEAD`

A bit like PHP functions, when git comes across a shorthand notation, it
needs to resolve it. The order of resolution is defined here:
https://git-scm.com/docs/gitrevisions#Documentation/gitrevisions.txt-refnameegmasterheadsmasterrefsheadsmaster

The downstream code, i.e. the code that uses the base branch works just
fine with the full path. Likewise, the user passes the default base
branch, he can use either form.

So, in the end, there is not really any reason for us to further modify
the output of the git command.